datatalksclub-zoomcamp-2026/01-Containerization-and-IaC/pyproject.toml
[project]
name = "01-containerization-and-iac"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.3.1",
    "pandas>=3.0.0",
    "polars>=1.37.1",
    "psycopg2-binary>=2.9.11",
    "pyarrow>=23.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.46",
    "tqdm>=4.67.1",
]
